fix: update based on reviewers' suggestions
