+++ frontend/src/app/(homepage)/page.tsx
@@ -1,46 +1,13 @@
-"use client";
-
-import { useQuery } from "@tanstack/react-query";
-import { motion } from "framer-motion";
 import { Loader2Icon } from "lucide-react";
 import Image from "next/image";
 import Link from "next/link";
 import type React from "react";
-import { type ComponentProps } from "react";
-import { twMerge } from "tailwind-merge";
+import { Suspense } from "react";
 
+import Block from "@/components/ui/block";
 import { buttonVariants } from "@/components/ui/button";
+import { createUsosService } from "@/lib/usos";
 import { cn } from "@/lib/utils";
-
-const Block = ({
-  className,
-  ...rest
-}: ComponentProps<(typeof motion)["div"]> & { className: string }) => {
-  return (
-    <motion.div
-      variants={{
-        initial: {
-          scale: 0.5,
-          y: 50,
-          opacity: 0,
-        },
-        animate: {
-          scale: 1,
-          y: 0,
-          opacity: 1,
-        },
-      }}
-      transition={{
-        type: "spring",
-        mass: 3,
-        stiffness: 400,
-        damping: 50,
-      }}
-      className={twMerge("col-span-4 rounded-lg p-6", className)}
-      {...rest}
-    />
-  );
-};
 
 const AnimationLogo = () => (
   <Block
@@ -76,76 +43,45 @@ const AnimationLogo = () => (
   </Block>
 );
 
-const JoinUsBlock = () => {
-  const query = useQuery({
-    queryKey: ["auth"],
-    queryFn: async () => {
-      const res = await fetch("/api/profile");
+const JoinUsBlock = async () => {
+  try {
+    const usos = await createUsosService();
+    await usos.getProfile();
 
-      if (res.status === 401) {
-        throw new Error("Unauthorized");
-      }
-
-      return res.json();
-    },
-    retry: false,
-    gcTime: 5000,
-  });
-
-  return (
-    <Block className="flex flex-col items-center justify-center gap-6 md:gap-10">
-      <div className="">
-        <h1 className="text-center text-4xl font-medium leading-tight md:text-left">
-          <span className="font-inter tracking-wide text-white animate-in">
-            Stwórz swój plan używając{" "}
-            <span className="font-bold uppercase">darmowego</span> zapisownika!
-          </span>
-        </h1>
-      </div>
-      <div className="">
-        <p className="text-center text-white md:mr-4 md:text-2xl">
-          Zaloguj się do platformy USOS i stwórz swój plan na semestr!
-        </p>
-      </div>
-      <div className="z-50">
-        {query.isLoading ? (
-          <div className="flex h-20 animate-spin items-center justify-center text-xl text-white">
-            <Loader2Icon size={32} />
-          </div>
-        ) : query.isError ? (
-          <Link
-            href="/api/login"
-            data-umami-event="Landing - Go to planning"
-            className={buttonVariants({
-              size: "lg",
-              variant: "outline",
-              className: cn(
-                "h-20 cursor-wait self-center border-4 text-xl transition-all duration-300 md:mt-0 md:p-7",
-                "cursor-pointer hover:bg-white hover:shadow-[0_0_5px_rgb(200,200,255),0_0_10px_rgb(164,200,255)]",
-              ),
-            })}
-          >
-            Zaloguj się
-          </Link>
-        ) : (
-          <Link
-            href="/plans"
-            data-umami-event="Landing - Go to planning"
-            className={buttonVariants({
-              size: "lg",
-              variant: "outline",
-              className: cn(
-                "h-20 cursor-wait self-center border-4 text-xl transition-all duration-300 md:mt-0 md:p-7",
-                "cursor-pointer hover:bg-white hover:shadow-[0_0_5px_rgb(200,200,255),0_0_10px_rgb(164,200,255)]",
-              ),
-            })}
-          >
-            Twoje plany
-          </Link>
-        )}
-      </div>
-    </Block>
-  );
+    return (
+      <Link
+        href="/plans"
+        data-umami-event="Landing - Go to planning"
+        className={buttonVariants({
+          size: "lg",
+          variant: "outline",
+          className: cn(
+            "h-20 cursor-wait self-center border-4 text-xl transition-all duration-300 md:mt-0 md:p-7",
+            "cursor-pointer hover:bg-white hover:shadow-[0_0_5px_rgb(200,200,255),0_0_10px_rgb(164,200,255)]",
+          ),
+        })}
+      >
+        Twoje plany
+      </Link>
+    );
+  } catch (e) {
+    return (
+      <Link
+        href="/api/login"
+        data-umami-event="Landing - Go to planning"
+        className={buttonVariants({
+          size: "lg",
+          variant: "outline",
+          className: cn(
+            "h-20 cursor-wait self-center border-4 text-xl transition-all duration-300 md:mt-0 md:p-7",
+            "cursor-pointer hover:bg-white hover:shadow-[0_0_5px_rgb(200,200,255),0_0_10px_rgb(164,200,255)]",
+          ),
+        })}
+      >
+        Zaloguj się
+      </Link>
+    );
+  }
 };
 
 export default function HomePage() {
@@ -157,7 +93,33 @@ export default function HomePage() {
         </div>
       </div>
       <section className="flex justify-center">
-        <JoinUsBlock />
+        <Block className="flex flex-col items-center justify-center gap-6 md:gap-10">
+          <div className="">
+            <h1 className="text-center text-4xl font-medium leading-tight md:text-left">
+              <span className="font-inter tracking-wide text-white animate-in">
+                Stwórz swój plan używając{" "}
+                <span className="font-bold uppercase">darmowego</span>{" "}
+                zapisownika!
+              </span>
+            </h1>
+          </div>
+          <div className="">
+            <p className="text-center text-white md:mr-4 md:text-2xl">
+              Zaloguj się do platformy USOS i stwórz swój plan na semestr!
+            </p>
+          </div>
+          <div className="z-50">
+            <Suspense
+              fallback={
+                <div className="flex h-20 animate-spin items-center justify-center text-xl text-white">
+                  <Loader2Icon size={32} />
+                </div>
+              }
+            >
+              <JoinUsBlock />
+            </Suspense>
+          </div>
+        </Block>
       </section>
     </div>
   );
