+++ frontend/src/app/plans/create/[id]/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import React from "react";
 
-import CreateNewPlanPage from "./_components/CreateNewPlanPage";
+import { CreateNewPlanPage } from "./_components/CreateNewPlanPage";
 
 interface PageProps {
   params: Promise<{ id: string }>;
@@ -14,7 +14,7 @@ export const metadata: Metadata = {
 
 export default async function CreateNewPlan({ params }: PageProps) {
   const { id } = await params;
-  if (!id || typeof id !== "string" || id.length === 0) {
+  if (typeof id !== "string" || id.length === 0) {
     return notFound();
   }
